KS-1-4 исправил тип переменных

+++ src/main/kotlin/lesson_1/lesson_1_task_4.kt
@@ -3,8 +3,8 @@ fun main() {
     val length: Long = 40_868_600_000
     val age: Byte = 27
     val partOfDay: Double = 0.075
-    val seconds: Int = 6480
-    val partOfYear: Double = 2.0547945205479453E-4
+    val seconds: Short = 6480
+    val partOfYear: Float = 2.0547945205479453E-4F
     val apogee: Int = 327000
 
     println("""
